+++ grok/Haskell/pythag-triples/app/pythagTriplesMain.hs
@@ -24,8 +24,8 @@ printTriples = mapM_ (putStrLn . showTriple)
 --   Each (a, b, c) have no common factors.
 --   a & c are always odd, b is always even 
 pythagTriplesFast :: Int -> Int -> [Triple]
-pythagTriplesFast start stop =
-  let n = stop
+pythagTriplesFast start end =
+  let n = end
       m = if start < 2
             then 2
             else start
@@ -42,8 +42,8 @@ pythagTriplesFast start stop =
 --   For each a, will find all corresponding b's and c'c 
 --   before moving onto the next a.
 pythagTriplesOrdered1:: Int -> Int -> [Triple]
-pythagTriplesOrdered1 start stop =
-  let n = stop
+pythagTriplesOrdered1 start end =
+  let n = end
       m = if start < 3
             then 3
             else start
@@ -59,8 +59,8 @@ pythagTriplesOrdered1 start stop =
 -- | Generate ordered Pythagorean Triples first by b then a
 --   Will only find a's where a < b
 pythagTriplesOrdered2 :: Int -> Int -> [Triple]
-pythagTriplesOrdered2 start stop = 
-  let n = stop
+pythagTriplesOrdered2 start end = 
+  let n = end
       m = if start < 4
             then 4
             else start
@@ -95,38 +95,46 @@ main :: IO ()
 main = do
   args <- getArgs
   case args of
-    ["-o1", numStr] -> printTriples $ pythagTriplesOrdered1 3 (read numStr)
-    ["-o2", numStr] -> printTriples $ pythagTriplesOrdered2 4 (read numStr)
-    ["-f", numStr]  -> printTriples $ pythagTriplesFast 2 (read numStr)
-    ["-fs", numStr] -> printTriples $ sort $ map sortTriple $ pythagTriplesFast 2 (read numStr)
+    ["-o1", start, end] -> printTriples $ pythagTriplesOrdered1 (read start) (read end)
+    ["-o1", end] -> printTriples $ pythagTriplesOrdered1 3 (read end)
+    ["-o2", start, end] -> printTriples $ pythagTriplesOrdered2 (read start) (read end)
+    ["-o2", end] -> printTriples $ pythagTriplesOrdered2 4 (read end)
+    ["-f", iter1, iter2]  -> printTriples $ pythagTriplesFast (read iter1) (read iter2)
+    ["-f", iters]  -> printTriples $ pythagTriplesFast 2 (read iters)
+    ["-fs", iter1, iter2] -> printTriples $ sort $ map sortTriple $ pythagTriplesFast (read iter1) (read iter2)
+    ["-fs", iters] -> printTriples $ sort $ map sortTriple $ pythagTriplesFast 2 (read iters)
     "-h":_          -> putStrLn $ usageString ++ infoString
     "--help":_      -> putStrLn $ usageString ++ infoString
-    "-o1":_         -> errorOut "option -o1 takes one argument"
-    "-o2":_         -> errorOut "option -o2 takes one argument"
-    "-f":_          -> errorOut "option -f takes one argument"
-    "-fs":_         -> errorOut "option -fs takes one argument"
+    "-o1":_         -> errorOut "option -o1 takes one or two arguments"
+    "-o2":_         -> errorOut "option -o2 takes one or two arguments"
+    "-f":_          -> errorOut "option -f takes one or two arguments"
+    "-fs":_         -> errorOut "option -fs takes one or two arguments"
     ('-':x:rest):_  -> errorOut $ '-':x:rest ++ " is an invalid option"
     [numStr]        -> printTriples $ pythagTriplesFast 2 (read numStr)
+    []              -> errorOut "called with no arguments"
     _               -> errorOut "called with invalid arguments"
   where
-    errorOut str =  error $ "\n  error: " ++ str
+    errorOut str =  error $ "\n  error, " ++ str
 
 usageString :: String
 usageString = unlines [
     "  "
-  , "  Usage: pythagTriples  [-o1|-o2|-f|-fs|-h] number"
+  , "  Usage: pythagTriples  [-o1|-o2|-f|-fs|-h] [begin] end"
   , "    where"
   , "      number = number of triples to print"
   , "    and"
   , "      -o1 Triples (a, b, c) are generated in lexiconical order,"
-  , "          that is a < b < c, where a,b,c have no common factors."
+  , "          that is a < b < c, where a,b,c have no common factors,"
+  , "          starting with a=3 (or begin) and endding with a=end."
   , "          Algorithm prints all possible b's and c's before"
   , "          going onto the next a."
   , "      -o2 Triples (a, b, c) are generated ordered first by b"
-  , "          then by a.  For each b, all a < b are generated"
+  , "          then by a.  For each b, all a < b are generated,"
+  , "          starting with b=4 (or begin) and endding with b=end."
   , "      -f  Use a fast algorithm where triples (a, b, c) are such"
   , "          that a is odd, b is even, and a,b,c have no common"
-  , "          factors."
+  , "          factors.  Both begin and end signify interation levels"
+  , "          used by the algorithm."
   , "      -fs Use above fast algorithm, sort results a < b < c."
   , "      -h  Print usage and general information."
   ]